requirements.txt
scrapy==2.11.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic
feedparser
redis==5.2.1
slack-bolt==1.22.0
boto3==1.39.4